scalafmt: split services to interface / Impl
